Extract join_and helper for easier re-use
